% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{alphabetacalc}
\alias{alphabetacalc}
\title{Calculate Alpha and Beta from Mean and Variance}
\usage{
alphabetacalc(mu, var)
}
\arguments{
\item{mu}{Mean.}

\item{var}{Variance.}
}
\value{
Numeric vector of alpha and beta.
}
\description{
Calculate Alpha and Beta from Mean and Variance
}
