% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{alphabetacalcvec}
\alias{alphabetacalcvec}
\title{Vector-based - Calculate Alpha and Beta from Mean and Variance}
\usage{
alphabetacalcvec(mu, var)
}
\arguments{
\item{mu}{Vector of mean.}

\item{var}{Vector of variance.}
}
\value{
Numeric matrix of alpha and beta.
}
\description{
Vector-based - Calculate Alpha and Beta from Mean and Variance
}
