% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{emstepB}
\alias{emstepB}
\title{Expectation maximization - Beta Distribution}
\usage{
emstepB(parmlist, xi, niter, epsilon, trunc, type = "free", glreturn = FALSE)
}
\arguments{
\item{parmlist}{A list containing initial alpha, mean, and variance values.}

\item{xi}{List of observations, in this case allele frequencies.}

\item{niter}{Max number of iterates.}

\item{epsilon}{Epsilon value for convergence tolerance. When the absolute delta log-likelihood is
below this value, convergence is reached.}

\item{trunc}{List of two values representing the lower and upper bounds, \eqn{c_{L}} and \eqn{c_{U}}.}

\item{type}{String indicating model type. Options: "free" (estimated parameter(s): alpha, mean, and variance), "fixed" (estimated parameter(s): alpha),
"fixed-2" (estimated parameter(s): alpha and variance), or "fixed-3" (estimated parameter(s): variance).
If avec is length of 1, fixed and fixed-3 will not be able to return a log-likelihood.}

\item{glreturn}{Logical statement, if \code{true}, the function will return the log-likelihood, negative log-likelihood, optimized parameter values, number of iterates, and the probability of each observation belonging to each mixture model. If \code{false}, the function will return the log-likelihood, negative log-likelihood, optimized parameter values, and number of iterates.}
}
\value{
List of elements including the log likelihood, the negative log likelihood, the number of iterates,
and the optimized parameter values. If
}
\description{
This function calculates the log-likelihood using
the expectation maximization algorithm with Nelder-Mead numerical optimization
and a beta distribution.
}
\examples{
  if(exists("crazy")){
    xi <- (xm[,2]/xm[,1])
    p = list(avec = c(0.11, 0.22, 0.34, 0.22, 0.11),
             mvec = c(0.20, 0.33, 0.50, 0.67, 0.80),
             svec = c(0.01, 0.01, 0.01, 0.01, 0.01));
    mout <- emstepB(p,
                    xi,
                    niter = 100,
                    epsilon = 0.1,
                    trunc = c(0.0,0.0))
}
}
