% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{emstepB3}
\alias{emstepB3}
\title{Expectation maximization - Beta + Beta + Beta Distribution}
\usage{
emstepB3(parmlist, xi, niter, epsilon, trunc)
}
\arguments{
\item{parmlist}{A list containing initial alpha, mean, and variance.}

\item{xi}{Matrix where the first column is total coverage and the second is the count of base A or B.}

\item{niter}{Max number of iterates.}

\item{epsilon}{Epsilon value for convergence tolerance. When the absolute delta log-likelihood is
below this value, convergence is reached.}

\item{trunc}{List of two values representing the lower and upper bounds,  \eqn{c_{L}} and \eqn{c_{U}}.}
}
\value{
List of elements including the negative log likelihood, the number of iterates,
and the optimized parameter values.
}
\description{
This function is made for the \code{Bclean()} function and preforms expectation maximization with Nelder-Mead
numerical optimization for beta distribution.
}
\examples{
 if(exists("crazy")){
    xi <- (xm[,2]/xm[,1])
    tcalc <- alphabetacalcvec(mu = c(0.287, 0.50, 0.713),
                              var = c(0.01, 0.01, 0.01))
    set <-  list(avec = c(0.25, 0.25, 0.25, 0.125, 0.125),
                 t1vec = c(tcalc[,1], 0.5, 0.33),
                 t2vec = c(tcalc[,2], 0.33, 0.5))
    checkB <- emstepB3(set,
                       xi,
                       1000,
                       0.1,
                       c(0,0))
}
}
