% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{emstepBB}
\alias{emstepBB}
\title{Expected maximization - Beta-Binomial Distribution}
\usage{
emstepBB(parmlist, xm, niter, epsilon, trunc, type = "free")
}
\arguments{
\item{parmlist}{A list containing initial alpha, mean, and variance.}

\item{xm}{Matrix where the first column is total coverage and the second is the count of base A or B.}

\item{niter}{Max number of iterates.}

\item{epsilon}{Epsilon value for convergence tolerance. When the absolute delta log-likelihood is
below this value, convergence is reached.}

\item{trunc}{List of two values representing the lower and upper bounds, $c_{L}$ and $c_{U}$.}

\item{type}{String indicating "Free" or "Fixed".}
}
\value{
List of elements including the negative log likelihood, the number of iterates,
and the optimized parameter values.
}
\description{
This function calculates the negative log-likelihood using
the expected maximization algorithm with Nelder-Mead numerical optimization
and beta-binomial distribution.
}
