% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{emstepNA}
\alias{emstepNA}
\title{Expected maximization - Normal Distribution}
\usage{
emstepNA(parmlist, xi, niter, epsilon, trunc, type = "free")
}
\arguments{
\item{parmlist}{A list containing initial alpha, mean, and variance values.}

\item{xi}{List of observations, in this case allele frequencies.}

\item{niter}{Max number of iterates.}

\item{epsilon}{Epsilon value for convergence tolerance. When the absolute delta log-likelihood is
below this value, convergence is reached.}

\item{trunc}{List of two values representing the lower and upper bounds, $c_{L}$ and $c_{U}$.}

\item{type}{String indicating model type. Options: "free" (estimated parameter(s): alpha, mean, and variance), "fixed" (estimated parameter(s): alpha),
"fixed_2" (estimated parameter(s): alpha and variance), or "fixed_3" (estimated parameter(s): variance).}
}
\value{
List of elements including the log-likelihood, the number of iterates,
and the optimized parameter values.
}
\description{
This function calculates the log-likelihood using
the expected maximization algorithm with the Normal Distribution.
This code is not identical to nQuire and uses an augmented likelihood.
}
