% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{emstepNU}
\alias{emstepNU}
\title{Expectation maximization - Normal and Uniform Distribution}
\usage{
emstepNU(parmlist, xi, niter, epsilon, trunc, type = "free", glreturn = FALSE)
}
\arguments{
\item{parmlist}{A list containing initial alpha, mean, and variance values. The list of alpha must include a proportion for the uniform mixture.}

\item{xi}{List of observations, in this case allele frequencies.}

\item{niter}{Max number of iterates.}

\item{epsilon}{Epsilon value for convergence tolerance. When the absolute delta log-likelihood is
below this value, convergence is reached.}

\item{trunc}{List of two values representing the lower and upper bounds, \eqn{c_{L}} and \eqn{c_{U}}.}

\item{type}{String indicating model type. Options: "free" (estimated parameter(s): alpha, mean, and variance), "fixed" (estimated parameter(s): alpha),
"fixed_2" (estimated parameter(s): alpha and variance), or "fixed_3" (estimated parameter(s): variance).}

\item{glreturn}{Logical statement, if \code{true}, the function will return the log-likelihood, negative log-likelihood, optimized parameter values, number of iterates, and the probability of each observation belonging to each mixture model. If \code{false}, the function will return the log-likelihood, negative log-likelihood, optimized parameter values, and number of iterates.}
}
\value{
List of elements including the log-likelihood, the number of iterates,
and the optimized parameter values.
}
\description{
This function calculates the log-likelihood using
the expectation maximization algorithm with the Normal-Uniform Distribution.
This code follows nQuire and does not use an augmented likelihood.
}
\examples{
 if(exists("crazy")){
  xi <- (xm[,2]/xm[,1])
  p = list(avec = c(0.11, 0.22, 0.34, 0.22, 0.11),
           mvec = c(0.20, 0.33, 0.50, 0.67, 0.80),
           svec = c(0.01, 0.01, 0.01, 0.01, 0.01));
  mout <- emstepNU(p,
                   xi,
                   niter = 100,
                   epsilon = 0.1,
                   trunc = c(0.0,0.0))
}
}
