% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muvarcalcvec.R
\name{muvarcalcvec}
\alias{muvarcalcvec}
\title{Variance calculation from Mean, Tau, and Sequencing Error}
\usage{
muvarcalcvec(mu, tau, error)
}
\arguments{
\item{mu}{Vector of means.}

\item{tau}{Sequence overdispersion parameter for read counts.}

\item{error}{Sequencing error rate.}
}
\value{
Mean and variance for the associated tau and error.
}
\description{
This function is used to calculate variance.
}
\examples{
var <- muvarcalcvec(mu = 0.5, tau = 0.01, error = 0.01)
}
