% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\docType{package}
\name{nQuack-package}
\alias{nQuack}
\alias{nQuack-package}
\title{nQuack: Predicting Ploidal Level from Sequence Data}
\description{
nQuack predicts ploidal level from sequence data using site-based heterozygosity and a mixture of models approach.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://mlgaynor.com/nQuack/}
  \item \url{https://github.com/mgaynor1/nQuack}
  \item Report bugs at \url{https://github.com/mgaynor1/nQuack/issue}
}

}
\author{
\strong{Maintainer}: Michelle L. Gaynor \email{shellyleegaynor@gmail.com} (\href{https://orcid.org/0000-0002-3912-6079}{ORCID}) [copyright holder]

}
\keyword{internal}
