% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nQuire_reformat}
\alias{nQuire_reformat}
\title{Data Preparation - Use nQuire's Data}
\usage{
nQuire_reformat(xm)
}
\arguments{
\item{xm}{A matrix with three columns: Total Coverage, Counts for Allele A, and Counts for Allele B.}
}
\value{
Numeric Matrix with total coverage and coverage for a randomly sampled allele.
}
\description{
This function reduce a three column data frame to
two columns by randomly sampling allele A or B for every site. This is used in our function \code{process_nquire()}
}
