% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare data - Step 1}
\usage{
prepare_data(name, inpath, outpath, tempfolder = "temp")
}
\arguments{
\item{name}{File name without the suffix. For example, if your file is called "frog.bam", this input should be "frog".}

\item{inpath}{Location of input file.}

\item{outpath}{Location for output file.}

\item{tempfolder}{Location for temp folder.}
}
\value{
Writes text file with the following columns: chromosome, position, depth, A, C, G, and T.
}
\description{
This function transforms a BAM file into a text file.
Specifically, this function uses \href{http://www.htslib.org/doc/samtools-mpileup.html}{samtools mpileup}
to translate your BAM into a tab-separated file. We then filter this file to remove indels and deletions.
When running this function, a temporary folder will be created (named 'temp/'), however this folder will be
removed once the process is complete.
}
\details{
Warning, due to the processing time needed for samtools mpileup,
this step may take some time. This function also requires samtools to be located locally. Please see
our \href{mlgaynor.com/nQuack/DataPreparation.html}{Data Preparation} article for more information. Warning, this writes a temporary folder
titled 'temp'. If you want to run multiple samples at once, we suggest you set the working directory to separate locations to ensure that
your temp folder/files are not overwritten.
}
