% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{process_data}
\alias{process_data}
\title{Process data - Step 2}
\usage{
process_data(
  file,
  min.depth = 2,
  max.depth.quantile.prob = 0.9,
  error = 0.01,
  trunc = c(0, 0)
)
}
\arguments{
\item{file}{Output txt file created with \code{prepare_data()}.}

\item{min.depth}{Minimum sequencing depth, default as 2.}

\item{max.depth.quantile.prob}{Maximum sequencing depth quantile cut off, default = 0.9.}

\item{error}{Sequencing error rate. If an \code{error} is provided,
sites will be retained where allele coverage is greater than the sequencing error rate times
the total coverage, but less than one minus the sequencing error rate times the total coverage.}

\item{trunc}{List of two values representing the lower and upper bounds, \eqn{c_{L}} and \eqn{c_{U}} which are used to filter allele frequencies.}
}
\value{
Numeric matrix with total coverage and coverage for a randomly sampled allele.
}
\description{
Based on the file generated with \code{prepare_data()}, which contains the total depth and sequencing coverage for each
nucleotide (A, C, G, and T), this function remove all but single nucelotide polymorphisms.
When supplied, this function will filter on coverage or  allele frequency. To filter by total coverage,
a user must supply the \code{min.depth} and \code{max.depth.quantile.prob}. If an \code{error} is provided,
sites will be retained where allele coverage is greater than the sequencing error rate times
the total coverage, but less than one minus the sequencing error rate times the total coverage.
Lastly, based on \code{trunc}, allele frequencies will be filtered based on a provided lower
and upper bound. Finally, the function samples a single allele frequency per site to avoid data duplication.
}
