% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{process_rcpp}
\alias{process_rcpp}
\title{Data Preparation - Matrix Filtering}
\usage{
process_rcpp(x, mindepth, maxprob, trunc, error)
}
\arguments{
\item{x}{Matrix with five columns: Depth, A, C, G, and T.}

\item{mindepth}{Minimum depth, default = 15.}

\item{maxprob}{Maximum depth quantile cut off, default = 0.9.}

\item{trunc}{List of two values representing the lower and upper bounds, $c_{L}$ and $c_{U}$.}

\item{error}{Sequencing error rate.}
}
\value{
Numeric Matrix with total coverage and coverage for a randomly sampled allele.
}
\description{
Based on supplied matrix with total depth and sequencing coverage for each
nucleotide (A, C, G, and T) this function remove all but single nucelotide polymorphisms.
When supplied, this function will filter on coverage or  allele frequency. Finally, the function
samples a single allele frequency per site to avoid data duplication.
}
