% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quackBetaBinom.R
\name{quackBetaBinom}
\alias{quackBetaBinom}
\title{Model Selection - Expected Maximization - Beta-Binomial Mixture}
\usage{
quackBetaBinom(
  xm,
  samplename,
  cores,
  parallel = FALSE,
  trunc = c(0, 0),
  lowvar = FALSE,
  tau = NA,
  error = NA,
  free = FALSE
)
}
\arguments{
\item{xm}{Matrix with two columns with total coverage and coverage
for a randomly sampled allele.}

\item{samplename}{Name of sample to be included in output.}

\item{cores}{Threads available to run process in parallel.}

\item{parallel}{default = FALSE, set to true if cores > 1.}

\item{trunc}{List of two values representing the lower and upper bounds for
allele frequency truncation , \eqn{c_{L}} and \eqn{c_{U}}. If allele frequency
truncation was done to remove error, then you do not need to truncate
the expected. If no truncation has been done, this should be set to c(0,0),
which is the default.}

\item{lowvar}{Default to FALSE. When false, variance is equal to 0.01.
If set to TRUE and tau and error are not provided,
the variance will be set as 0.001.}

\item{tau}{Sequencing overdispersion parameter. If tau and error are provided,
the variance of each mixture will be inferred from these values.
If not, the variance by default is equal to 0.01 or 0.001.}

\item{error}{Sequencing error rate. If tau and error are provided,
the variance of each mixture will be inferred from these values. If not, the variance by default is equal to 0.01 or 0.001.}

\item{free}{default = FALSE, skip the free model calculation and does not
calculate delta log-likelihood.}
}
\value{
BIC scores and log-likelihood (LL) mixture models including diploid,
triploid, tetraploid, pentaploid, and hexaploid. When free = TRUE,
the delta log-likelihood (dLL) is calculated based on
the associated free model (without or with a uniform mixture).
For BIC or delta-log likelihood, the smallest score is the most likely model.
For LL, the largest score is the most likely model.
}
\description{
This function uses the expected maximization of both
the beta-binomial and beta-binomial-uniform mixture models for model selection.
Here we can run up to 32 mixture models.
}
