% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quackit.R
\name{quackit}
\alias{quackit}
\title{Model Selection - Based on BIC or Log-Likelihood}
\usage{
quackit(
  model_out,
  summary_statistic = "BIC",
  mixtures = c("diploid", "triploid", "tetraploid", "hexaploid", "pentaploid")
)
}
\arguments{
\item{model_out}{Data frame containing, at minimum, columns labeled LL, type, mixture, distribution, and BIC.}

\item{summary_statistic}{May be equal to BIC or LL.}

\item{mixtures}{Defaults to \code{c("diploid", "triploid", "tetraploid", "hexaploid", "pentaploid")}.}
}
\value{
Returns data frame with the most likely model for each set of mixtures.
Includes the best and second best mixtures, as well as the difference between the two.
We only use BIC or LL to compare within each distribution and type.
To identify the most accurate model, you will need to compare accuracy across distributions
and types using a set of known samples. The distributions include
Normal, Beta, and Beta-Binomial - each with and without a uniform mixture.
The type indicates which parameters are estimated for the mixtures:
all parameters (\code{type = 'free'}, only used to calculate delta log-likelihood),
only alpha  (\code{type = 'fixed'}), only alpha and variance (\code{type = 'fixed_2'}),
and only variance (\verb{type ='fixed_3}) to be estimated for each mixture.
}
\description{
This function is for model interpretation.
}
