% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{resample_xm}
\alias{resample_xm}
\title{Calculate Alpha and Beta from Mean and Variance}
\usage{
resample_xm(xm, n)
}
\arguments{
\item{xm}{Matrix with total coverage and coverage at a randomly sampled allele.}

\item{n}{Length of matrix.}
}
\value{
Randomly sampled matrix.
}
\description{
Calculate Alpha and Beta from Mean and Variance
}
