% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setconvert.R
\name{setconvert}
\alias{setconvert}
\title{Calculate Variance from Mean, Tau, and Sequencing Error}
\usage{
setconvert(set, tau, error)
}
\arguments{
\item{set}{A list of lists, each of the lists must contain avec, mvec, and svec.}

\item{tau}{Sequence overdispersion parameter for read counts.}

\item{error}{Sequencing error rate.}
}
\value{
Mean and variance for the associated tau and error.
}
\description{
This function is used to replace variance in mixture model sets.
}
\examples{
 set <- c()
 set[[1]] =  list(avec = c(1.00), mvec = c(0.50), svec = c(0.01));
 set[[2]] =  list(avec = c(0.50, 0.50), mvec = c(0.67, 0.33), svec = c(0.01, 0.01));
 exset <- setconvert(set, tau = 0.01, error = 0.001)
}
