% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xm.R
\docType{data}
\name{xm}
\alias{xm}
\title{Simulated data for examples
xm <- sim.ind.simple(mvec = c(0.5))}
\format{
A data frame with 2 columns and 50000 rows
}
\usage{
xm
}
\description{
Simulated data for examples
xm <- sim.ind.simple(mvec = c(0.5))
}
\keyword{internal}
