% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{emstepB3}
\alias{emstepB3}
\title{Expected maximization - Beta + Beta + Beta Distribution}
\usage{
emstepB3(parmlist, xi, niter, epsilon, trunc)
}
\arguments{
\item{parmlist}{A list containing initial alpha, mean, and variance.}

\item{xi}{Matrix where the first column is total coverage and the second is the count of base A or B.}

\item{niter}{Max number of iterates.}

\item{epsilon}{Epsilon value for convergence tolerance. When the absolute delta log-likelihood is
below this value, convergence is reached.}

\item{trunc}{List of two values representing the lower and upper bounds, $c_{L}$ and $c_{U}$.}
}
\value{
List of elements including the negative log likelihood, the number of iterates,
and the optimized parameter values.
}
\description{
This function is made for the \code{Bclean()} function and preforms expected maximization with Nelder-Mead
numerical optimization for beta distribution.
}
