% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{emstepBBU}
\alias{emstepBBU}
\title{Expectation maximization - Beta-Binomial and Uniform Distributions}
\usage{
emstepBBU(parmlist, xm, niter, epsilon, trunc, type = "free")
}
\arguments{
\item{parmlist}{A list containing initial alpha, mean, and variance values.}

\item{xm}{Matrix where the first column is total coverage and the second is the count of base A or B.}

\item{niter}{Max number of iterates.}

\item{epsilon}{Epsilon value for convergence tolerance. When the absolute delta log-likelihood is
below this value, convergence is reached.}

\item{trunc}{List of two values representing the lower and upper bounds, $c_{L}$ and $c_{U}$.}

\item{type}{String indicating model type. Options: "free" (estimated parameter(s): alpha, mean, and variance), "fixed" (estimated parameter(s): alpha),
"fixed-2" (estimated parameter(s): alpha and variance), or "fixed-3" (estimated parameter(s): variance).
If avec is length of 1, fixed and fixed-3 will not be able to return a log-likelihood.}
}
\value{
List of elements including the log likelihood, the negative log likelihood, the number of iterates,
and the optimized parameter values.
}
\description{
This function calculates the log-likelihood using
the expectation-maximization algorithm with Nelder-Mead numerical optimization
and beta distribution with one uniform mixture.
}
