% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{estepB3}
\alias{estepB3}
\title{E-Step for Expectation Maximization - Beta + Beta + Beta Distribution}
\usage{
estepB3(parmlist, xi, trunc)
}
\arguments{
\item{parmlist}{A list containing initial alpha, mean, and variance.}

\item{xi}{List of observations, in this case allele frequencies.}

\item{trunc}{List of two values representing the lower and upper bounds, $c_{L}$ and $c_{U}$.}
}
\description{
This is used in the \code{Bclean()} function. Here we complete
the E-Step and calculate the log-likelihood. Modifications include a correction for
the truncated distribution.
}
