% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quackit.R
\name{quackit}
\alias{quackit}
\title{Model Selection - Based on BIC or Log-Likelihood}
\usage{
quackit(
  model_out,
  summary_statistic = "BIC",
  mixtures = c("diploid", "triploid", "tetraploid", "hexaploid", "pentaploid")
)
}
\arguments{
\item{model_out}{Data frame containing, at minimum, columns labeled LL, type, mixture, distribution, and BIC.}

\item{summary_statistic}{May be equal to BIC or LL.}

\item{mixtures}{Defaults to \code{c("diploid", "triploid", "tetraploid", "hexaploid", "pentaploid")}.}
}
\value{
Returns data frame with the most likely model for each set of mixtures. Includes the best and second best mixtures, as well as the difference between the two.
}
\description{
This function is for model interpretation.
}
