% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setconvert.R
\name{setconvert}
\alias{setconvert}
\title{Calculate Variance from Mean, Tau, and Sequencing Error}
\usage{
setconvert(set, tau, error)
}
\arguments{
\item{set}{A list of lists, each of the lists must contain avec, mvec, and svec.}

\item{tau}{Sequence overdispersion parameter for read counts.}

\item{error}{Sequencing error rate.}
}
\value{
Mean and variance for the associated tau and error.
}
\description{
This function is used to replace variance in mixture model sets.
}
